# docs/swagger.yaml
definitions:
  dto.CredentialSignInDto:
    properties:
      deviceToken:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.CustomerSignup:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  dto.DriverSignup:
    type: object
  dto.InsertBookingTransfer:
    properties:
      addDropoffPoint:
        type: integer
      addPickupPoint:
        type: integer
      adultSeater:
        type: integer
      carModelId:
        type: string
      childSeater:
        type: integer
      fromLatCoordinate:
        type: number
      fromLngCoordinate:
        type: number
      fromLocation:
        type: string
      grandTotal:
        type: number
      id:
        type: string
      passengerNotes:
        type: string
      pickUpDate:
        type: string
      price:
        type: number
      toLatCoordinate:
        type: number
      toLngCoordinate:
        type: number
      toLocation:
        type: string
      totalDistance:
        type: number
    type: object
  dto.InsertBookingTransferRating:
    properties:
      bookingTransferAssignedID:
        type: string
      carManagementId:
        type: string
      customerId:
        type: string
      rating:
        type: integer
    type: object
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      kind:
        type: string
      message:
        type: string
    type: object
  dto.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UserSignup:
    properties:
      address:
        type: string
      createdBy:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
info:
  contact: {}
paths:
  /auth/c/signin:
    post:
      consumes:
      - application/json
      description: A Customer authentication sign-in method
      parameters:
      - description: customer login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Customer Sign-In
      tags:
      - Authentication - Customer
  /auth/c/signup:
    post:
      consumes:
      - application/json
      description: Customer Sign-Up
      parameters:
      - description: customer registration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CustomerSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Customer Sign-Up
      tags:
      - Authentication - Customer
  /auth/d/signin:
    post:
      consumes:
      - application/json
      description: A Driver authentication sign-in method for mobile apps
      parameters:
      - description: driver login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Driver Sign-In
      tags:
      - Authentication - Driver
  /auth/d/signup:
    post:
      consumes:
      - application/json
      description: Driver Sign-Up
      parameters:
      - description: driver registration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DriverSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Driver Sign-Up
      tags:
      - Authentication - Driver
  /auth/i/signin:
    post:
      consumes:
      - application/json
      description: A Internal authentication sign-in method
      parameters:
      - description: internal login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CredentialSignInDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Internal Sign-In
      tags:
      - Authentication - Internal
  /auth/i/signup:
    post:
      consumes:
      - application/json
      description: Internal Sign-Up
      parameters:
      - description: internal registration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Internal Sign-Up
      tags:
      - Authentication - Internal
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Universal Reset Password
      parameters:
      - description: universal reset password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DriverSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Universal Reset Password
      tags:
      - Authentication - Reset Password
  /booking-assigned/transfer:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/accept/{id}:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to Accept specific booking transfer by
        it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to accept booking transfer
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/accepted:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
...
  /booking-assigned/transfer/cancelled:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/completed:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view booking
        transfer that assigned and whose are not accepted by the driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Booking Transfer Assigned
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/ongoing:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to set Complete Pickup specific booking
        transfer by it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set status to Complete PickUp and notify passenger
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/ongoing/{id}:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to set status to be "OnGoing" to specific
        booking transfer by it's ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set status to ongoing and notify passenger that driver on
        it's way to pickup
      tags:
      - Booking Assgined - Transfer
  /booking-assigned/transfer/switch/{id}/{driverId}/{plateNumber}:
    get:
      consumes:
      - application/json
      description: A POST Request for Driver to set Complete Pickup specific booking
        transfer by it's ID
      parameters:
      - description: BookingTransferAssigned ID
        in: path
        name: id
        required: true
        type: string
      - description: DriverID ID
        in: path
        name: driverId
        required: true
        type: string
      - description: Car Plate Number
        in: path
        name: plateNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set status to Complete PickUp and notify passenger
      tags:
      - Booking Assgined - Transfer
  /booking/transfer:
    post:
      consumes:
      - application/json
      description: A POST Request to create a new record for costumer to make a new
        Transfer booking to the driver
      parameters:
      - description: insert booking transfer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.InsertBookingTransfer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to create booking transfer request by customer
      tags:
      - Booking - Transfer
  /booking/transfer/all:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for customer to view booking
        transfer that already created
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to get all booking transfer made by customer
      tags:
      - Booking - Transfer
  /booking/transfer/cancel/{id}:
    post:
      consumes:
      - application/json
      description: A POST Request to cancel the selected booking transfer and notify
        the assigned driver
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking Transfer Assigned ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to cancel booking transfer by customer
      tags:
      - Booking - Transfer
  /booking/transfer/rating:
    post:
      consumes:
      - application/json
      description: A POST Request to give a driver's car a rating based on completed
        booking
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: insert booking transfer rating
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.InsertBookingTransferRating'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to give rating to the driver car
      tags:
      - Booking Transfer - Rating
  /car-management/create:
    post:
      consumes:
      - application/json
      description: A POST Request to create a car and assigned to the driver whose
        creating it
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: |-
          FrontCarPhoto     *multipart.FileHeader `json:"frontCarPhoto" swaggerignore:"true"`
          LicensePhoto      *multipart.FileHeader `json:"licensePhoto" swaggerignore:"true"`
          RoadTaxPhoto      *multipart.FileHeader `json:"roadTaxPhoto" swaggerignore:"true"`
          VEPPhoto          *multipart.FileHeader `json:"vepPhoto" swaggerignore:"true"`
        in: formData
        name: carManagementType
        required: true
        type: string
      - in: formData
        name: carModelId
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: plateNumber
        required: true
        type: string
      - description: front car photo upload
        in: formData
        name: frontCarPhoto
        required: true
        type: file
      - description: license photo upload
        in: formData
        name: licensePhoto
        required: true
        type: file
      - description: road tax photo upload
        in: formData
        name: roadTaxPhoto
        required: true
        type: file
      - description: vep photo upload
        in: formData
        name: vepPhoto
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to create or assign a car to the driver
      tags:
      - Car Management
  /car-management/delete/{id}:
    delete:
      consumes:
      - application/json
      description: A DELETE Request to delete record car management
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to delete car management made by driver
      tags:
      - Car Management
  /car-management/get:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records for driver to view list of
        added car management that already created
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to get all car management made by driver
      tags:
      - Car Management
  /car-management/set/active/{id}:
    put:
      consumes:
      - application/json
      description: A UPDATE Request to set record car management to active
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set active car management made by driver
      tags:
      - Car Management
  /car-management/set/inactive/{id}:
    put:
      consumes:
      - application/json
      description: A UPDATE Request to set record car management to inactive
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to set inactive car management made by driver
      tags:
      - Car Management
  /car-management/update/{id}:
    put:
      consumes:
      - application/json
      description: A UPDATE Request to update record car management
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: carModelId
        required: true
        type: string
      - in: formData
        name: licensePhoto
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: plateNumber
        required: true
        type: string
      - description: front car photo upload
        in: formData
        name: frontCarPhoto
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to update car management made by driver
      tags:
      - Car Management
  /company/create:
    post:
      consumes:
      - application/json
      description: A POST Request to create a company
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: companyAddress
        required: true
        type: string
      - in: formData
        name: companyCountry
        required: true
        type: string
      - in: formData
        name: companyName
        required: true
        type: string
      - in: formData
        name: companyNumber
        required: true
        type: string
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: phoneNumber
        required: true
        type: string
      - description: company certificate upload
        in: formData
        name: companyCertificate
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to create company
      tags:
      - Company
  /company/create/approval/{id}:
    get:
      consumes:
      - application/json
      description: A POST Request to accept company submission request
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to Approve company registration
      tags:
      - Company
  /company/driver/all/{id}:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a all records
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to get all drivers by company id
      tags:
      - Company
  /driver-balance/get:
    get:
      consumes:
      - application/json
      description: A GET Request to fetch a record by driver to check their wallet
        information
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Method to get wallet information by driver
      tags:
      - Driver Balance
...
swagger: "2.0"